moved types into their own folders/packages
