+++ .github/workflows/ci.yml
@@ -30,4 +30,4 @@ jobs:
           go-version: ${{ matrix.go-version }}
       - name: Code
         uses: actions/checkout@v3
-      - run: go test -v
+      - run: go get . && go test -v
