+++ move.go
@@ -3,14 +3,14 @@ package pokego
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/mazylol/pokego/types/moves"
 	"log"
 
 	"github.com/mazylol/pokego/cache"
-	"github.com/mazylol/pokego/types"
 )
 
 // GetMove returns a Move struct containing information about the Move with the given name.
-func GetMove(name string) (types.Move, error) {
+func GetMove(name string) (moves.Move, error) {
 	move, err := cache.GetMoveFromCache(name)
 	if err != nil {
 		body, err := callApi(fmt.Sprintf("move/%v", name))
@@ -32,7 +32,7 @@ func GetMove(name string) (types.Move, error) {
 }
 
 // GetMoveList returns a list of Move names. You have to include a limit for the amount of names you want.
-func GetMoveList(limit int) (types.MoveList, error) {
+func GetMoveList(limit int) (moves.MoveList, error) {
 	moveList, err := cache.GetMoveListFromCache(limit)
 	if err != nil {
 		body, err := callApi(fmt.Sprintf("move?limit=%v", limit))
